+++ mycollab-services/src/main/java/com/mycollab/module/user/service/mybatis/UserServiceDBImpl.java
@@ -191,7 +191,7 @@ public class UserServiceDBImpl extends DefaultService<String, User, UserSearchCr
         if ((record.getEmail()) != null && !record.getUsername().equals(record.getEmail())) {
             UserExample ex = new UserExample();
             ex.createCriteria().andUsernameEqualTo(record.getEmail());
-            int numUsers = userMapper.countByExample(ex);
+            Long numUsers = userMapper.countByExample(ex);
             if (numUsers > 0) {
                 throw new UserInvalidInputException(String.format("Email %s is already existed in system. Please choose another email.",
                         record.getEmail()));
@@ -223,7 +223,7 @@ public class UserServiceDBImpl extends DefaultService<String, User, UserSearchCr
         if (!record.getUsername().equals(record.getEmail())) {
             UserExample ex = new UserExample();
             ex.createCriteria().andUsernameEqualTo(record.getEmail());
-            int numUsers = userMapper.countByExample(ex);
+            Long numUsers = userMapper.countByExample(ex);
             if (numUsers > 0) {
                 throw new UserInvalidInputException(String.format("Email %s is already existed in system. Please choose another email.",
                         record.getEmail()));
